replacing old implementation with redesigned components
