+++ packages/stepper/src/Stepper.story.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import times from 'lodash/times';
 import { storiesOf } from '@storybook/react';
 import { number } from '@storybook/addon-knobs';
-import { Step, Stepper } from '.';
-import NewStepper from './NewStepper';
+import Stepper from '.';
 import { addons } from '@storybook/addons';
 import { CHANGE } from '@storybook/addon-knobs';
 
@@ -14,6 +13,7 @@ storiesOf('Stepper', module)
     <div style={{ width: 1000 }}>
       <Stepper
         currentStep={number('Step', 0, { min: 0, max: 6 })}
+        completedStepsShown={number('Completed steps shown', 2, { min: 1 })}
         maxDisplayedSteps={5}
       >
         <div>Overview</div>
@@ -26,7 +26,7 @@ storiesOf('Stepper', module)
       </Stepper>
     </div>
   ))
-  .add('Redesign', () => {
+  .add('Many', () => {
     let currentStep = number('Step', 1, { min: 0 });
     const numSteps = number('Number of steps', 10, { min: 1 });
     const maxDisplayedSteps = number('Max displayed', 5, { min: 1 });
@@ -43,39 +43,13 @@ storiesOf('Stepper', module)
 
     return (
       <div style={{ width: 1000 }}>
-        <NewStepper
+        <Stepper
           currentStep={currentStep}
           completedStepsShown={completedStepsShown}
           maxDisplayedSteps={maxDisplayedSteps}
         >
           {times(numSteps, count => (
-            <Step key={count}>Step {count + 1}</Step>
-          ))}
-        </NewStepper>
-      </div>
-    );
-  })
-  .add('Many', () => {
-    let currentStep = number('Step', 0, { min: 0 });
-    const numSteps = number('Number of steps', 7, { min: 1 });
-
-    // Can't dynamically change the max, so we manually enforce it
-    if (currentStep + 1 > numSteps) {
-      channel.emit(CHANGE, {
-        name: 'Step',
-        value: numSteps - 1,
-      });
-      currentStep = numSteps - 1;
-    }
-
-    return (
-      <div style={{ width: 1000 }}>
-        <Stepper
-          currentStep={currentStep}
-          maxDisplayedSteps={number('Max displayed', 5, { min: 1 })}
-        >
-          {times(numSteps, count => (
-            <Step key={count}>Step {count + 1}</Step>
+            <div key={count}>Step {count + 1}</div>
           ))}
         </Stepper>
       </div>
