+++ packages/stepper/src/index.ts
@@ -1,1 +1,3 @@
-export { Step, Stepper } from './Stepper';
+import Stepper from './Stepper';
+export { Step } from './Step';
+export default Stepper;
